New component changes and css
Removed WeatherData component and created a new component WeatherAPI which takes in parameters and handles a request to backend

Also added new Component DisplayData that displays the query results in transparent dark border box shadow with close button. Requres to add eventListener to close on enter.

Added new state to searchBar which takes in boolean if user hits enter or not to display the query results.

In addition added more ES6 Syntax like props deconstructor