New components added and  funtional component change
Added new file ImageSlider that changes the image of the background when on hover.

Changed Background class and NavBar from functional to class components to pass in props and have a state for background image.

New functions added to update background of image when on hover and onMouseClick event when user hovers over navbar divs to change background image.